Navbar styling and responsiveness
Added a smaller version of the logo for mobile. Navbar now only displays the bigger logo on screen sizes of 600 pixels wide or larger.

Wrapped navbar in new withResizeHandler component (in hoc folder) so it updates on window resize. It is an edge case, but it would look pretty bad after resizing to a smaller window with the bigger logo.

Added appropriate buttons to both the logged in and logged out user experience. Only the login and sign up buttons are functional as of right now.

Styled the navbar appropriately according to mocks. Should also be responsive to screen size.